Fix unimplemented methods under build constraints
