+++ lib/teleterm/autoupdate/service_notwindows.go
@@ -1,0 +1,33 @@
+// Teleport
+// Copyright (C) 2026 Gravitational, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+//go:build !windows
+
+package autoupdate
+
+import (
+	"context"
+
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+
+	api "github.com/gravitational/teleport/gen/proto/go/teleport/lib/teleterm/auto_update/v1"
+)
+
+// GetInstallationMetadata implements [api.AutoUpdateServiceServer].
+func (*Service) GetInstallationMetadata(context.Context, *api.GetInstallationMetadataRequest) (*api.GetInstallationMetadataResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetInstallationMetadata not implemented")
+}
