+++ lib/teleterm/vnet/service.go
@@ -27,8 +27,6 @@ import (
 
 	"github.com/gravitational/trace"
 	"github.com/jonboulle/clockwork"
-	"google.golang.org/grpc/codes"
-	grpcstatus "google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/gravitational/teleport"
@@ -832,13 +830,3 @@ func (r *disabledTelemetryUsageReporter) ReportSSHSession(profileName, rootClust
 }
 
 func (r *disabledTelemetryUsageReporter) Stop() {}
-
-// CheckInstallTimeRequirements implements [api.VnetServiceServer].
-func (*Service) CheckInstallTimeRequirements(context.Context, *api.CheckInstallTimeRequirementsRequest) (*api.CheckInstallTimeRequirementsResponse, error) {
-	return nil, grpcstatus.Errorf(codes.Unimplemented, "method CheckInstallTimeRequirements not implemented")
-}
-
-// GetBackgroundItemStatus implements [api.VnetServiceServer].
-func (*Service) GetBackgroundItemStatus(context.Context, *api.GetBackgroundItemStatusRequest) (*api.GetBackgroundItemStatusResponse, error) {
-	return nil, grpcstatus.Errorf(codes.Unimplemented, "method GetBackgroundItemStatus not implemented")
-}
